requirements.txt
# CUDA Version : 11.3 (with cuDNN 8.2.1)
# PyTorch
--find-links https://download.pytorch.org/whl/cu113/torch_stable.html
torch==1.10.0+cu113
torchvision==0.11.1+cu113

pyyaml==6.0
lmdb==1.2.1
tensorboard==2.7.0
# lpips==0.1.4
